Separate files for contract subcommands
