+++ commands/contract/contractCommands.js
@@ -1,0 +1,22 @@
+const { SlashCommandBuilder, SlashCommandSubcommandBuilder} = require("discord.js");
+const { handleContractCreate } = require("./contract_create");
+
+module.exports = {
+    data: new SlashCommandBuilder()
+        .setName("contract")
+        .setDescription("Feedback contract commands")
+        
+        .addSubcommand(new SlashCommandSubcommandBuilder()
+            .setName("create")
+            .setDescription("Creates a feedback contract")
+        ),
+
+    async execute(interaction) {
+        switch (interaction.options.getSubcommand()) {
+            case ("create"):
+                handleContractCreate(interaction);
+            default:
+                return;
+        }
+    },
+};
